Made changes to file
Notable changes:

line 54: changed cd 00-learning-styles/notes to cd 00-learning-style/notes

lines 78-80: added songs/ to the filepath so the files would be created within the songs folder

line 161: The command on this line does not work. I asked a couple of people about it and they said it wasn't a valid command. I'd recommend getting rid of it and having the student rename the files manually to continue in the tutorial.

line 179: The student is already in the card-games folder, so to delete all of the files beginning with game in that folder we just need to call rm hearts/game*

I hope these notes helped. Please let me know if anything is incorrect. Thanks for letting me work on this!