Update string-slicing-concatenation-worksheet.md
I was working through the exercises and numbers 10 and 12 didn't have an opening ". In irb it changed the prompt to have the " in it. I googled what that meant and was able to find the solution on stackoverflow.

It didn't seem like finding missing " was part of the exercise so I thought I'd fix it for the next person. If it was intentional please disregard this pull request.