New video link in Programming Grammar readme
The old link was to a video we had hosted on Vimeo. Since we've now
moved away from Vimeo to Panopto, I've updated the link to a video on
that platform.

I _believe_ these are the same video, but it's pretty hard to confirm
because the original link was actually to our Vimeo channel as a whole,
rather than a specific video.